UPD: Refactor download trailers into multiple files
